pyproject.toml
[tool.poetry]
name = "ksem-transpiler"
version = "0.1.0"
description = ""
authors = ["Nymphana <hi@nymphana.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
attrs = "^23.2.0"
pydantic = "^2.7.1"
bidict = "^0.23.1"
ruamel-yaml = "^0.18.6"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ipython = "^8.24.0"
mypy = "^1.10.0"
black = "^24.4.2"
ruff = "^0.4.5"
pre-commit = "^3.7.1"
pyright = "^1.1.364"

[tool.black]
skip-magic-trailing-comma = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
